requirements.txt
dbt-core==1.8.9
dbt-postgres==1.8.2
pandas==2.3.3
psycopg2-binary==2.9.11
python-dotenv==1.2.1